Run global checks and validate description when label is added to PR
Currently, global check workflows do not trigger when a bot user makes a commit to the PR. This happens in flow repo quite a lot as we have some bots committing to auto fix lint issues and for dummy commits to trigger e2e runs.

Since global checks do not run, the PR gets stuck and requires a manual dummy commit and running the e2e tests again. 

Adding a trigger for global checks on label added to manually trigger the workflow without need for a commit and running CCI tests again.